Add files via upload
I have made some changes based on the review. the class names are not the same as written in the review it would be a problem for me to change all those classes. please check the classes and you will find that it contains the things you want. Thank you.